Move combination renderers test to orca test suite
