+++ plotly/tests/test_io/test_renderers.py
@@ -253,44 +253,6 @@ def test_browser_renderer_show(fig1, renderer):
     assert_not_requirejs(html)
 
 
-# Combination
-# -----------
-def test_mimetype_combination(fig1):
-    pio.renderers.default = 'pdf+jupyterlab'
-
-    # Configure renderer so that we can use the same parameters
-    # to build expected image below
-    pio.renderers['pdf'].width = 400
-    pio.renderers['pdf'].height = 500
-    pio.renderers['pdf'].scale = 1
-
-    # pdf
-    image_bytes = pio.to_image(
-        fig1, format='pdf', width=400, height=500, scale=1)
-
-    image_str = base64.b64encode(image_bytes).decode('utf8')
-
-    # plotly mimetype
-    plotly_mimetype_dict = json.loads(
-        pio.to_json(fig1, remove_uids=False))
-
-    plotly_mimetype_dict['config'] = {
-        'plotlyServerURL': 'https://plot.ly'}
-
-    # Build expected bundle
-    expected = {
-        'application/pdf': image_str,
-        plotly_mimetype: plotly_mimetype_dict,
-    }
-
-    pio.renderers.render_on_display = False
-    assert fig1._repr_mimebundle_(None, None) is None
-
-    pio.renderers.render_on_display = True
-    bundle = fig1._repr_mimebundle_(None, None)
-    assert bundle == expected
-
-
 # Validation
 # ----------
 @pytest.mark.parametrize(
